% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapajos.R
\docType{data}
\name{tapajos}
\alias{tapajos}
\title{Tapajós National Forest}
\format{
An object of class 'sfc_POLYGON' within the bounding box 
xmin: -55.41127 ymin: -4.114584 
xmax: -54.7973 ymax: -2.751706
}
\source{
The data was provided by the Chico Mendes Institute for Biodiversity 
Conservation (ICMBio) via \url{https://www.protectedplanet.net/en}
}
\usage{
data(tapajos)
}
\description{
Geometries for the Tapajós National Forest, a protected 
area in the Brazilian Amazon
}
\keyword{datasets}
