% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.geojson.R
\name{as.geojson}
\alias{as.geojson}
\alias{as.geojson.default}
\alias{as.geojson.sf}
\title{Methods to coerce geographical coordinates into a geojson polygon}
\usage{
as.geojson(lonlat, dist = 0.00001, nQuadSegs = 2L, ...)

\method{as.geojson}{default}(lonlat, dist = 0.00001, nQuadSegs = 2L, ...)

\method{as.geojson}{sf}(lonlat, dist = 0.00001, nQuadSegs = 2L, ...)
}
\arguments{
\item{lonlat}{a \code{data.frame} or matrix with geographical coordinates
'lonlat', in that order, or an object of class \code{\link[sf]{sf}} with
geometry type 'POINT' or 'POLYGON'}

\item{dist}{numeric, buffer distance for all \code{lonlat}}

\item{nQuadSegs}{integer, number of segments per quadrant}

\item{...}{further arguments passed to \code{\link[sf]{sf}} methods}
}
\value{
An object of class 'geosjon' for each row in \code{lonlat}
}
\description{
Take single points from geographical coordinates and coerce into a geojson of
 geometry 'Polygon'
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Default S3 Method
# random geographic points within bbox(10, 12, 45, 47)
library("sf")

set.seed(123)
lonlat <- data.frame(lon = runif(1, 10, 12),
                     lat = runif(1, 45, 47))

gjson <- as.geojson(lonlat)

#################

# S3 Method for objects of class 'sf'
# random geographic points within bbox(10, 12, 45, 47)
library("sf")

set.seed(123)
lonlat <- data.frame(lon = runif(5, 10, 12),
                     lat = runif(5, 45, 47))

lonlat <- st_as_sf(lonlat, coords = c("lon","lat"))

gjson <- as.geojson(lonlat)
\dontshow{\}) # examplesIf}
}
\concept{utility functions}
