% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\docType{package}
\name{chirps}
\alias{chirps}
\alias{chirps-package}
\title{API Client for CHIRPS}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

API Client for the Climate Hazards Group InfraRed Precipitation
  with Station Data 'CHIRPS'. The 'CHIRPS' data is a 35+ year quasi-global
  rainfall data set, which incorporates 0.05 arc-degrees resolution satellite
  imagery, and in-situ station data to create gridded rainfall time series for
  trend analysis and seasonal drought monitoring. For more details on 'CHIRPS'
  data please visit its official home page <https://www.chc.ucsb.edu/data/chirps>.
  Requests from large time series (> 10 years) and large geographic coverage 
  (global scale) may take several minutes.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: 
 \url{https://github.com/agrobioinfoservices/chirps}}
\item{Static documentation: 
 \url{https://agrobioinfoservices.github.io/chirps/}}
\item{Report bugs: 
 \url{https://github.com/agrobioinfoservices/chirps/issues}}
}
}
\author{
Kauê de Sousa and Adam H. Sparks and William Ashmall and Jacob van
 Etten and Svein Ø. Solberg
}
