% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downCHIRPS.R
\name{downCHIRPS}
\alias{downCHIRPS}
\title{Download/update CHIRPS v2}
\usage{
downCHIRPS(
  path = getwd(),
  product,
  time_span,
  res,
  format = "tif",
  crop_by,
  ...
)
}
\arguments{
\item{path}{}

\item{product}{character. Type of product to download from \url{ftp://ftp.chg.ucsb.edu/pub/org/chg/products/CHIRPS-2.0/}
\tabular{ll}{
Product \tab Description \cr
global_2-monthly \tab Quasi-global 2-monthly fields \cr
global_3-monthly \tab Quasi-global 3-monthly fields \cr
global_annual \tab Quasi-global annual fields \cr
global_daily  \tab Quasi-global daily fields \cr
global_dekad  \tab Quasi-global dekadal fields \cr
global_monthly \tab Quasi-global monthly fields \cr
global_pentad \tab Quasi-global pentadal fields.
}}

\item{time_span}{vector with two dates, \code{c(start,end)}}

\item{res}{numeric. Resolution for daily product. It could be .05 or .25.}

\item{format}{character. 'tif' or 'gz'}

\item{crop_by}{character vector of two values. Provided to download a
cropped data for either continent or country}
}
\value{

}
\description{
For downloading global satellite estimates of precipitation from porduct CHIRPS v2
}
\examples{

#Download ten days of the global_daily product
#downCHIRPS(path=getwd(), product = 'global_daily', time_span = c('2019-01-01','2019-01-10'),res=.05)

}
