% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chirps.R
\name{get_chirps}
\alias{get_chirps}
\title{Get CHIRPS}
\usage{
get_chirps(lonlat, dates, operation = 5)
}
\arguments{
\item{lonlat}{a data.frame of geographic coordinates of longitude and latitude in that order}

\item{dates}{a character of start and end dates in that order in the format YYYY-MM-DD}

\item{operation}{optional, an integer that represents which type of statistical operation to perform on the dataset}
}
\value{
A data frame of CHIRPS data including latlon and dates
}
\description{
Fetch the Climate Hazards Group InfraRed Precipitation with Station Data via 
ClimateSERV API Client
}
\details{
Types of operation supported by ClimateSERV are,
max = 0, min = 1, median = 2, range = 3, sum = 4, average = 5
}
\examples{
lonlat <- data.frame(lon = runif(10, 5, 7),
                     lat = runif(10, 60, 62))
dates <- c("2019-10-01","2019-11-01")

get_chirps(lonlat, dates)
}
\seealso{
\url{https://climateserv.servirglobal.net} and \url{http://chg.geog.ucsb.edu/data/chirps/}
}
