% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chirps.R
\name{get_chirps}
\alias{get_chirps}
\title{Get CHIRPS precipitation data}
\usage{
get_chirps(lonlat, dates, operation = 5)
}
\arguments{
\item{lonlat}{a data.frame of geographic coordinates of longitude and latitude in that order}

\item{dates}{a character of start and end dates in that order in the format YYYY-MM-DD}

\item{operation}{optional, an integer that represents which type of statistical operation 
to perform on the dataset}
}
\value{
A data frame of CHIRPS data including:
\item{id}{the index for the rows in \code{lonlat}}
\item{dates}{the dates from which CHIRPS was requested}
\item{lon}{the longitude as provided in \code{lonlat}}
\item{lat}{the latitude as provided in \code{lonlat}}
\item{chirps}{the CHIRPS value in mm}
}
\description{
Fetch the Climate Hazards Group InfraRed Precipitation with Station Data via 
ClimateSERV API Client. ClimateSERV works with geojson strings of type 'Polygon'. The 
input 'lonlat' are then transformed into polygons with a small buffer area around 
the point.
}
\details{
Types of operation supported by ClimateSERV are,
max = 0, min = 1, median = 2, sum = 4, average = 5
}
\examples{
\donttest{

# Two random points across the Tapajos National Forest in Brazil
set.seed(123)
lonlat <- data.frame(lon = runif(2, -55, -54),
                     lat = runif(2, -3, -2.7))

dates <- c("2017-12-15","2018-01-20")


df <- get_chirps(lonlat, dates)
} 
}
\references{
Funk C. et al. (2015). Scientific Data, 2, 150066. \url{https://doi.org/10.1038/sdata.2015.66}
}
\seealso{
ClimateSERV <https://climateserv.servirglobal.net>
}
