% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esi.R
\name{get_esi}
\alias{get_esi}
\alias{get_esi.default}
\alias{get_esi.sf}
\title{Get evaporative stress index (ESI) data}
\usage{
get_esi(object, dates, operation = 5, period = 1, ...)

\method{get_esi}{default}(object, dates, operation = 5, period = 1, ...)

\method{get_esi}{sf}(object, dates, operation = 5, period = 1, as.sf = FALSE, ...)
}
\arguments{
\item{object}{input, an object of class \code{\link{data.frame}} or
\code{\link{sf}}}

\item{dates}{a character of start and end dates in that order in the format
YYYY-MM-DD}

\item{operation}{optional, an integer that represents which type of
statistical operation to perform on the dataset}

\item{period}{an integer value for the period of ESI data, four weeks period
= 1, twelve weeks = 2}

\item{...}{further arguments passed to \code{\link{sf}} methods. See details}

\item{as.sf}{logical, returns an object of class \code{sf} for S3 method of
class \code{\link{sf}}}
}
\value{
A data frame of ESI data:
\item{id}{the index for the rows in \code{object}}
\item{dates}{the dates from which ESI was requested}
\item{lon}{the longitude as provided in \code{object}}
\item{lat}{the latitude as provided in \code{object}}
\item{esi}{the ESI value}
}
\description{
Get evaporative stress index (ESI) from SERVIR Global
via ClimateSERV API Client. ESI is available every four (or twelve) weeks
 from 2001 to present.
The dataset may contain cloudy data which is returned as NAs.
ClimateSERV works with geojson of type 'Polygon'. The input \code{object} is
 then transformed into polygons with a small buffer area around the point.
}
\details{
operation: supported operations are max = 0, min = 1, median = 2, sum = 4,
 average = 5

dist: numeric, buffer distance for each \code{object} coordinate

nQuadSegs: integer, number of segments per buffer quadrant
}
\examples{
\donttest{
library("chirps")

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

dates <- c("2017-12-15","2018-06-20")

# by default the function set a very small buffer around the points
# which can return NAs due to cloudiness in ESI data

dat <- get_esi(lonlat, dates = dates)

dat

# the argument dist passed through sf increase the buffer area

dat <- get_esi(lonlat, dates = dates, dist = 0.1)

dat
} 
}
\references{
ClimateSERV <https://climateserv.servirglobal.net>

Evaporative Stress Index <http://catalogue.servirglobal.net/Product?product_id=198>
}
