% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esi.R
\name{get_esi}
\alias{get_esi}
\alias{get_esi.default}
\alias{get_esi.sf}
\alias{get_esi.geojson}
\title{Get evaporative stress index (ESI) data}
\usage{
get_esi(object, dates, operation = 5, period = 1, ...)

\method{get_esi}{default}(object, dates, operation = 5, period = 1, ...)

\method{get_esi}{sf}(object, dates, operation = 5, period = 1, as.sf = FALSE, ...)

\method{get_esi}{geojson}(object, dates, operation = 5, period = 1, as.geojson = FALSE, ...)
}
\arguments{
\item{object}{input, an object of class \code{\link[base]{data.frame}} (or
any other object that can be coerced to data.frame), \code{geojson} or 
\code{\link[sf]{sf}}}

\item{dates}{a character of start and end dates in that order in the format
"YYYY-MM-DD"}

\item{operation}{optional, an integer that represents which type of
statistical operation to perform on the dataset}

\item{period}{an integer value for the period of ESI data, 
four weeks period = 1, twelve weeks = 2}

\item{...}{further arguments passed to \code{\link[sf]{sf}} methods
See details}

\item{as.sf}{logical, returns an object of class \code{\link[sf]{sf}}}

\item{as.geojson}{logical, returns an object of class \code{geojson}}
}
\value{
A data frame of \acronym{ESI} data:
\item{id}{the index for the rows in \code{object}}
\item{dates}{the dates from which ESI was requested}
\item{lon}{the longitude as provided in \code{object}}
\item{lat}{the latitude as provided in \code{object}}
\item{esi}{the ESI value}
}
\description{
Get evaporative stress index (\acronym{ESI}) from \acronym{SERVIR} Global
via ClimateSERV \acronym{API} Client. \acronym{ESI} is available every four
 (or twelve) weeks from 2001 to present.
The dataset may contain cloudy data which is returned as \code{NA}s.
ClimateSERV works with geojson of type 'Polygon'. The input \code{object} is
 then transformed into polygons with a small buffer area around the point.
}
\details{
\bold{operation}: supported operations are:  
 \tabular{rll}{
 \bold{operation}      \tab    \tab \bold{value}\cr
 max                   \tab =  \tab 0\cr
 min                   \tab =  \tab 1\cr
 median                \tab =  \tab 2\cr
 sum                   \tab =  \tab 4\cr
 average               \tab =  \tab 5 (\emph{default value})\cr
 }

\bold{dist}: numeric, buffer distance for each \code{object} coordinate

\bold{nQuadSegs}: integer, number of segments per buffer quadrant
}
\note{
get_esi may return some warning messages given by 
\code{\link[sf]{sf}}, please look sf documentation for 
possible issues.
}
\examples{
\donttest{

lonlat <- data.frame(lon = c(-55.0281,-54.9857),
                     lat = c(-2.8094, -2.8756))

dates <- c("2017-12-15","2018-06-20")

# by default the function set a very small buffer around the points
# which can return NAs due to cloudiness in ESI data

dt <- get_esi(lonlat, dates = dates)

# the argument dist passed through sf increase the buffer area

dt <- get_esi(lonlat, dates = dates, dist = 0.1)


############################################

# S3 method for objects of class 'sf'
library("sf")

## geometry 'POINT'
example("tapajos", package = "chirps")

dates <- c("2017-11-15", "2017-12-31")

dt <- get_esi(tp_point, dates, dist = 0.1)

## geometry 'POLYGON'
p1 <- matrix(c(10.67, 49.90,
               10.57, 49.80,
               10.47, 49.90,
               10.57, 50.00,
               10.67, 49.90),
             nrow = 5, ncol = 2, byrow = TRUE)

p1 <- st_polygon(list(p1))

p2 <- matrix(c(11.67, 45.59,
               11.57, 45.49,
               11.47, 45.59,
               11.57, 45.69,
               11.67, 45.59),
             nrow = 5, ncol = 2, byrow = TRUE)

p2 <- st_polygon(list(p2))

pol <- data.frame(x = c(1, 2))
pol$geometry = st_sfc(p1, p2)

pol <- st_as_sf(pol)

dt <- get_esi(pol, dates = c("2018-01-01", "2018-02-20"))

} 
}
\references{
ClimateSERV \url{https://climateserv.servirglobal.net}
}
