% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapajos.R
\docType{data}
\name{tapajos}
\alias{tapajos}
\title{Tapajos National Forest}
\format{
An object of class 'sfc_POLYGON' within the bounding box 
xmin: -55.41127 ymin: -4.114584 
xmax: -54.7973 ymax: -2.751706
}
\source{
The data was provided by the Chico Mendes Institute via
\url{https://www.protectedplanet.net/}.
}
\usage{
tapajos
}
\description{
Geometries for the Tapajos National Forest, a protected 
area in the Brazilian 
Amazon \url{http://www.icmbio.gov.br/flonatapajos/}
}
\examples{
\donttest{
library("sf")

data("tapajos", package = "chirps")

# sample three points within the Tapajos area
set.seed(1234)
tp_point <- st_sample(tapajos, 3)

# coerce as sf points
tp_point <- st_as_sf(tp_point)
}
}
\keyword{datasets}
